+++ src/main/scala/LSH/hashFunctions/Hyperplane.scala
@@ -10,14 +10,14 @@ import scala.util.Random
   */
 class Hyperplane(k:Int) extends HashFunction(k) {
 
-  var hyperPlanes = new ArrayBuffer[Vector[Double]]
+  var hyperPlanes = new ArrayBuffer[Vector[Float]]
 
   // Initialize k random hyperplanes for H[i]
   for(m <- 0 until k) {
     hyperPlanes += generateRandomV(4096, System.currentTimeMillis())
   }
 
-  def apply(v: Vector[Double]): String = {
+  def apply(v: Vector[Float]): String = {
     var sb : StringBuilder = new StringBuilder
     for(h <- hyperPlanes) {
       sb ++= hash(v,h).toString
@@ -25,13 +25,13 @@ class Hyperplane(k:Int) extends HashFunction(k) {
     // TODO Dont use string
     sb.toString()
   }
-  def hash(v: Vector[Double], randomV: Vector[Double]): Int = {
+  def hash(v: Vector[Float], randomV: Vector[Float]): Int = {
     if (Distance.dotProduct(v, randomV) > 0) 1 else 0
   }
 
   // TODO check the seq structure
-  def generateRandomV(size: Int, seed: Long): Vector[Double] = {
-    val buf = new ArrayBuffer[Double]
+  def generateRandomV(size: Int, seed: Long): Vector[Float] = {
+    val buf = new ArrayBuffer[Float]
     val rnd = new Random(seed)
     for (i <- 0 until size)
       buf += (if (rnd.nextGaussian() < 0) -1 else 1)
